requirements.txt
Django>=4.2
djangorestframework>=3.14

beautifulsoup4>=4.12
cloudscraper>=1.2.71
requests>=2.31
